Organize mixins into various packages
