+++ src/main/java/jard/alchym/mixin/rendering/PlayerAnimMixin$1.java
@@ -1,4 +1,4 @@
-package jard.alchym.mixin;
+package jard.alchym.mixin.rendering;
 
 import jard.alchym.client.ExtraPlayerDataAccess;
 import net.minecraft.entity.EntityType;
@@ -11,11 +11,18 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
+/***
+ *  PlayerAnimMixin$1
+ *  Adds an additional previous velocity variable to player entities, which are then ticked at the beginning
+ *  of every movement logic tick, to permit smooth interpolation of the velocity for rendering purposes.
+ *
+ *  Created by jard at 19:32 on January, 02, 2021.
+ ***/
 @Mixin (PlayerEntity.class)
-public abstract class PlayerAnimMixin1 extends LivingEntity implements ExtraPlayerDataAccess {
+public abstract class PlayerAnimMixin$1 extends LivingEntity implements ExtraPlayerDataAccess {
     Vec3d previousVel = Vec3d.ZERO;
 
-    protected PlayerAnimMixin1 (EntityType<? extends LivingEntity> entityType, World world) {
+    protected PlayerAnimMixin$1 (EntityType<? extends LivingEntity> entityType, World world) {
         super (entityType, world);
     }
