+++ src/main/java/jard/alchym/mixin/stacklimit/BumpStackLimitMixin$2.java
@@ -1,4 +1,4 @@
-package jard.alchym.mixin;
+package jard.alchym.mixin.stacklimit;
 
 import jard.alchym.AlchymReference;
 import net.minecraft.entity.Entity;
