+++ src/main/resources/alchym.mixins.json
@@ -3,19 +3,20 @@
   "package": "jard.alchym.mixin",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
-    "BumpStackLimitMixin$1",
-    "BumpStackLimitMixin$2",
     "FluidSolubleMixin",
-    "InventoryMixin",
     "ItemStackMixin",
-    "RevolverHitsoundMixin"
+    "RevolverHitsoundMixin",
+
+    "stacklimit.BumpStackLimitMixin$1",
+    "stacklimit.BumpStackLimitMixin$2",
+    "stacklimit.InventoryMixin"
   ],
   "client": [
-    "EnableStencilMixin",
-    "MatrixStackMixin",
-    "PlayerAnimMixin1",
-    "PlayerAnimMixin2",
-    "PlayerAnimMixin3"
+    "rendering.EnableStencilMixin",
+    "rendering.MatrixStackMixin",
+    "rendering.PlayerAnimMixin$1",
+    "rendering.PlayerAnimMixin$2",
+    "rendering.PlayerAnimMixin$3"
   ],
   "injectors": {
     "defaultRequire": 1
